refactor: extract text_normalize, input_hooks, capture_profile into own modules
