+++ plugin/plugins/galgame_plugin/ocr_runtime_types.py
@@ -107,6 +107,9 @@ except ImportError:  # pragma: no cover
 
 from typing import ClassVar
 
+from .ocr_text_normalize import *
+from .ocr_capture_profile import *
+
 
 __all__ = [
     "DetectedGameWindow",
@@ -656,180 +659,6 @@ def _build_window_key(*, process_name: str, pid: int, hwnd: int, title: str) -> 
     return f"ocrwin:{digest}"
 
 
-def _looks_like_self_window_title(title: str) -> bool:
-    normalized = _normalize_window_title(title)
-    return any(token in normalized for token in _SELF_WINDOW_TITLE_SUBSTRINGS)
-
-
-def _looks_like_self_window_path(exe_path: str) -> bool:
-    lowered = str(exe_path or "").strip().lower()
-    if not lowered:
-        return False
-    return any(token in lowered for token in _SELF_WINDOW_PATH_SUBSTRINGS)
-
-
-def _looks_like_self_ui_text(text: str) -> bool:
-    normalized = normalize_text(text).strip().lower()
-    if not normalized:
-        return False
-    return any(token in normalized for token in _SELF_UI_GUARD_SUBSTRINGS)
-
-
-def _looks_like_english_overlay_label(line: str) -> bool:
-    words = re.findall(r"[a-z]+", normalize_text(line).strip().lower())
-    if not words:
-        return False
-    if any(word not in _ENGLISH_GAME_OVERLAY_WORDS for word in words):
-        return False
-    return True
-
-
-def _looks_like_non_english_overlay_label(line: str) -> bool:
-    compact = re.sub(
-        r"[\s\-_.,，。:：;；!！?？/\\|()\[\]【】「」『』]+",
-        "",
-        normalize_text(line).strip().lower(),
-    )
-    if not compact:
-        return False
-    remainder = compact
-    matched = False
-    for token in sorted(_NON_ENGLISH_GAME_OVERLAY_SUBSTRINGS, key=len, reverse=True):
-        if token in remainder:
-            matched = True
-            remainder = remainder.replace(token, "")
-    return matched and not remainder
-
-
-def _looks_like_game_overlay_normalized_text(normalized: str) -> bool:
-    normalized = str(normalized or "").strip()
-    if not normalized:
-        return False
-    lowered = normalized.lower()
-    lines = _stripped_ocr_lines(lowered)
-    non_english_overlay_lines = [
-        line for line in lines if _looks_like_non_english_overlay_label(line)
-    ]
-    if bool(lines) and len(non_english_overlay_lines) == len(lines):
-        return True
-    english_overlay_lines = [
-        line for line in lines if _looks_like_english_overlay_label(line)
-    ]
-    return bool(lines) and len(english_overlay_lines) == len(lines)
-
-
-def _looks_like_game_overlay_text(text: str) -> bool:
-    normalized = normalize_text(text).strip().lower()
-    return _looks_like_game_overlay_normalized_text(normalized)
-
-
-def _coerce_prefixed_choice_lines(lines: list[str]) -> list[str]:
-    if len(lines) < 2:
-        return []
-    choices: list[str] = []
-    for line in lines:
-        match = _MENU_PREFIX_RE.match(line)
-        if match is None:
-            return []
-        text = match.group(1).strip()
-        if not text:
-            return []
-        choices.append(text)
-    return choices
-
-
-def _looks_like_dialogue_line(text: str) -> bool:
-    normalized = normalize_text(text).strip()
-    if not normalized:
-        return False
-    return any(marker in normalized for marker in _DIALOGUE_LINE_MARKERS)
-
-
-def _looks_like_ocr_dialogue_text(text: str) -> bool:
-    normalized = normalize_text(text).replace("\n", " ").strip()
-    return _looks_like_ocr_dialogue_normalized_text(normalized)
-
-
-def _looks_like_ocr_dialogue_normalized_text(normalized: str) -> bool:
-    normalized = str(normalized or "").replace("\n", " ").strip()
-    if not normalized:
-        return False
-    significant_chars = _significant_char_count(normalized)
-    if significant_chars < 2 or significant_chars > 220:
-        return False
-    if _OCR_DIALOGUE_STRONG_PUNCTUATION_RE.search(normalized):
-        return True
-    if _OCR_DIALOGUE_WEAK_PUNCTUATION_RE.search(normalized) and significant_chars >= 8:
-        return True
-    return False
-
-
-def _clean_ocr_dialogue_text(text: str) -> str:
-    normalized = normalize_text(text).replace("\n", " ").replace("　", "").strip()
-    if not normalized:
-        return ""
-    cleaned = normalized
-    cleaned = _OCR_TRAILING_GARBAGE_AFTER_SENTENCE_RE.sub(r"\1", cleaned).strip()
-    if _significant_char_count(cleaned) >= 10:
-        cleaned = _OCR_TRAILING_ORPHAN_AFTER_SENTENCE_RE.sub(r"\1", cleaned).strip()
-    cleaned = _OCR_TRAILING_GARBAGE_AFTER_BRACKET_RE.sub(r"\1", cleaned).strip()
-    cleaned = _OCR_TRAILING_GARBAGE_AFTER_DASH_RE.sub(r"\1", cleaned).strip()
-    return cleaned
-
-
-def _drop_ocr_chrome_noise_lines(text: str, *, window_title: str = "") -> str:
-    lines = [line.strip() for line in str(text or "").splitlines()]
-    meaningful = [line for line in lines if line]
-    if len(meaningful) < 2:
-        return str(text or "")
-    filtered = [
-        line
-        for line in meaningful
-        if not _looks_like_temperature_status_line(line)
-        and not _looks_like_window_title_line(line, window_title)
-    ]
-    if filtered and len(filtered) < len(meaningful):
-        return "\n".join(filtered)
-    return str(text or "")
-
-
-def _ocr_stability_key(text: str) -> str:
-    normalized = normalize_text(str(text or "")).replace("\n", " ").strip().lower()
-    if not normalized:
-        return ""
-    return _OCR_STABILITY_IGNORED_CHARS_RE.sub("", normalized)
-
-
-def _ocr_stability_keys_match(left: str, right: str) -> bool:
-    left_key = str(left or "")
-    right_key = str(right or "")
-    if not left_key or not right_key:
-        return False
-    if left_key == right_key:
-        return True
-    if len(left_key) == len(right_key) and len(left_key) >= 8:
-        distance = sum(1 for left_char, right_char in zip(left_key, right_key) if left_char != right_char)
-        allowed_distance = max(1, int(len(left_key) * 0.08))
-        return distance <= allowed_distance
-    return False
-
-
-def _prefer_ocr_stability_text(existing: str, current: str) -> str:
-    existing_text = normalize_text(str(existing or "")).strip()
-    current_text = normalize_text(str(current or "")).strip()
-    if not existing_text:
-        return current_text
-    if not current_text:
-        return existing_text
-    existing_has_strong_end = bool(_OCR_DIALOGUE_STRONG_PUNCTUATION_RE.search(existing_text[-2:]))
-    current_has_strong_end = bool(_OCR_DIALOGUE_STRONG_PUNCTUATION_RE.search(current_text[-2:]))
-    if existing_has_strong_end != current_has_strong_end:
-        return existing_text if existing_has_strong_end else current_text
-    if _significant_char_count(current_text) > _significant_char_count(existing_text):
-        return current_text
-    return existing_text
-
-
 def _rapidocr_runtime_cache_key(
     *,
     install_target_dir_raw: str,
@@ -881,35 +710,6 @@ def _store_rapidocr_runtime_cache(
 ) -> None:
     _prune_rapidocr_runtime_cache(now)
     _RAPIDOCR_RUNTIME_CACHE[key] = (runtime, now)
-
-
-def _coerce_plain_choice_lines(lines: list[str]) -> list[str]:
-    if not 2 <= len(lines) <= _AIHONG_MENU_MAX_LINES:
-        return []
-    choices: list[str] = []
-    seen: set[str] = set()
-    for line in lines:
-        text = normalize_text(str(line or "")).replace("\n", " ").strip()
-        if not text or _looks_like_dialogue_line(text):
-            return []
-        if _significant_char_count(text) > _AIHONG_MENU_MAX_SIGNIFICANT_CHARS:
-            return []
-        if text in seen:
-            continue
-        seen.add(text)
-        choices.append(text)
-    if not 2 <= len(choices) <= _AIHONG_MENU_MAX_LINES:
-        return []
-    return choices
-
-
-def _coerce_choice_lines(lines: list[str], *, allow_plain_text: bool = False) -> list[str]:
-    choices = _coerce_prefixed_choice_lines(lines)
-    if choices:
-        return choices
-    if allow_plain_text:
-        return _coerce_plain_choice_lines(lines)
-    return []
 
 
 def _aihong_choice_boxes(
@@ -1045,168 +845,6 @@ def _frame_choice_bounds_metadata(frame: Any, *, text_source: str = "") -> dict[
     return metadata
 
 
-@dataclass(slots=True)
-class OcrCaptureProfile:
-    left_inset_ratio: float = DEFAULT_OCR_CAPTURE_LEFT_INSET_RATIO
-    right_inset_ratio: float = DEFAULT_OCR_CAPTURE_RIGHT_INSET_RATIO
-    top_ratio: float = DEFAULT_OCR_CAPTURE_TOP_RATIO
-    bottom_inset_ratio: float = DEFAULT_OCR_CAPTURE_BOTTOM_INSET_RATIO
-
-    def to_dict(self) -> dict[str, float]:
-        return {
-            "left_inset_ratio": self.left_inset_ratio,
-            "right_inset_ratio": self.right_inset_ratio,
-            "top_ratio": self.top_ratio,
-            "bottom_inset_ratio": self.bottom_inset_ratio,
-        }
-
-    @classmethod
-    def from_dict(cls, value: dict[str, Any]) -> OcrCaptureProfile:
-        return cls(
-            left_inset_ratio=float(
-                value.get("left_inset_ratio", DEFAULT_OCR_CAPTURE_LEFT_INSET_RATIO)
-            ),
-            right_inset_ratio=float(
-                value.get("right_inset_ratio", DEFAULT_OCR_CAPTURE_RIGHT_INSET_RATIO)
-            ),
-            top_ratio=float(value.get("top_ratio", DEFAULT_OCR_CAPTURE_TOP_RATIO)),
-            bottom_inset_ratio=float(
-                value.get("bottom_inset_ratio", DEFAULT_OCR_CAPTURE_BOTTOM_INSET_RATIO)
-            ),
-        )
-
-
-def _score_ocr_text(text: str) -> tuple[float, int, int]:
-    normalized = normalize_text(text)
-    if not normalized:
-        return (-1.0, 0, 0)
-    cjk_count = len(_CJK_CHAR_RE.findall(normalized))
-    kana_count = len(_KANA_CHAR_RE.findall(normalized))
-    ascii_tokens = _ASCII_TOKEN_RE.findall(normalized)
-    isolated_ascii_tokens = sum(
-        1 for token in ascii_tokens if len(token) == 1 and token.lower() not in {"i", "a"}
-    )
-    multi_char_ascii_tokens = sum(1 for token in ascii_tokens if len(token) > 1)
-    significant_chars = sum(1 for ch in normalized if not ch.isspace())
-    score = (
-        (cjk_count * 5.0)
-        + (kana_count * 4.0)
-        + (multi_char_ascii_tokens * 1.5)
-        + (significant_chars * 0.2)
-        - (isolated_ascii_tokens * 2.0)
-    )
-    return (score, cjk_count + kana_count, significant_chars)
-
-
-_PUNCTUATION_CONFUSION_FIXES = [
-    (re.compile(r"(?<=[^\x00-\x7F])\.(?![\x00-\x7F])"), "。"),
-    (re.compile(r"(?<=[^\x00-\x7F])\s*,\s*(?=[^\x00-\x7F])"), "、"),
-    (re.compile(r"(?<=[^\x00-\x7F])!(?![\x00-\x7F])"), "！"),
-    (re.compile(r"(?<=[^\x00-\x7F])\?(?![\x00-\x7F])"), "？"),
-]
-
-
-def _fix_ocr_punctuation_confusion(text: str) -> str:
-    value = str(text or "")
-    for pattern, replacement in _PUNCTUATION_CONFUSION_FIXES:
-        value = pattern.sub(replacement, value)
-    return value
-
-
-def _significant_char_count(text: str) -> int:
-    return sum(1 for ch in str(text or "") if not ch.isspace())
-
-
-def _looks_like_noise_ocr_text(text: str) -> bool:
-    normalized = normalize_text(str(text or "")).strip()
-    return _looks_like_noise_normalized_text(normalized)
-
-
-def _looks_like_noise_normalized_text(normalized: str) -> bool:
-    normalized = str(normalized or "").strip()
-    if not normalized:
-        return True
-    significant_chars = _significant_char_count(normalized)
-    cjk_or_kana_count = len(_CJK_CHAR_RE.findall(normalized)) + len(_KANA_CHAR_RE.findall(normalized))
-    if cjk_or_kana_count <= 0 and significant_chars <= 2:
-        return True
-    return False
-
-
-def _classify_cjk_text(text: str) -> str:
-    """Return RapidOCR lang_type: japan, korean, ch, or unknown."""
-    if not text or not text.strip():
-        return "unknown"
-    if _HANGUL_RE.search(text):
-        return "korean"
-    if _HIRAGANA_RE.search(text) or _KATAKANA_RE.search(text):
-        return "japan"
-    if not _CJK_CHAR_RE.search(text):
-        return "unknown"
-
-    ja_votes = sum(1 for word in _JA_MARKER_WORDS if word in text)
-    zh_votes = sum(1 for word in _ZH_MARKER_WORDS if word in text)
-    if ja_votes > zh_votes:
-        return "japan"
-    if zh_votes > ja_votes:
-        return "ch"
-    if _KANA_BUD_RE.search(text):
-        return "japan"
-    return "ch"
-
-class _OcrLangDetector:
-    def __init__(self, window_size: int = 8, confirm_streak: int = 2) -> None:
-        self._window_size = max(1, int(window_size or 1))
-        self._confirm_streak = max(1, int(confirm_streak or 1))
-        self._buffer: list[str] = []
-        self._last_detected: str | None = None
-        self._streak = 0
-        self._switched_at: float | None = None
-
-    def feed(self, text: str) -> str | None:
-        cleaned = str(text or "").strip()
-        if not cleaned:
-            return None
-        if not (
-            _CJK_CHAR_RE.search(cleaned)
-            or _HIRAGANA_RE.search(cleaned)
-            or _KATAKANA_RE.search(cleaned)
-            or _HANGUL_RE.search(cleaned)
-        ):
-            return None
-
-        self._buffer.append(cleaned)
-        if len(self._buffer) < self._window_size:
-            return None
-
-        merged = " ".join(self._buffer)
-        self._buffer.clear()
-        detected = _classify_cjk_text(merged)
-        if detected == "unknown":
-            return None
-
-        if detected == self._last_detected:
-            self._streak += 1
-        else:
-            self._last_detected = detected
-            self._streak = 1
-
-        if self._streak >= self._confirm_streak:
-            return detected
-        return None
-
-    def reset(self, *, clear_switch_time: bool = False) -> None:
-        self._buffer.clear()
-        self._last_detected = None
-        self._streak = 0
-        if clear_switch_time:
-            self._switched_at = None
-
-    @property
-    def last_switched_at(self) -> float | None:
-        return self._switched_at
-
-
 def _prepare_ocr_image(image: Any, *, apply_filters: bool = True) -> Any:
     from PIL import Image, ImageFilter, ImageOps
 
@@ -1268,68 +906,6 @@ def _rapidocr_points(box: Any) -> list[tuple[float, float]]:
         except (TypeError, ValueError):
             continue
     return points
-
-
-def _should_insert_ascii_space(previous_text: str, next_text: str) -> bool:
-    if not previous_text or not next_text:
-        return False
-    left = previous_text[-1]
-    right = next_text[0]
-    return left.isascii() and right.isascii() and left.isalnum() and right.isalnum()
-
-
-def _join_ocr_segments(parts: list[str]) -> str:
-    rendered = ""
-    for part in parts:
-        normalized = normalize_text(str(part or "")).replace("\n", " ").strip()
-        if not normalized:
-            continue
-        if not rendered:
-            rendered = normalized
-            continue
-        if _should_insert_ascii_space(rendered, normalized):
-            rendered += " "
-        rendered += normalized
-    return rendered
-
-
-def _ocr_score_weight(text: str) -> int:
-    return max(_significant_char_count(text), 1)
-
-
-def _weighted_ocr_score(scores: Iterable[tuple[float, int]]) -> float:
-    total_weight = 0
-    weighted_sum = 0.0
-    for score, weight in scores:
-        normalized_weight = max(int(weight or 0), 1)
-        weighted_sum += float(score) * normalized_weight
-        total_weight += normalized_weight
-    if total_weight <= 0:
-        return 0.0
-    return weighted_sum / total_weight
-
-
-def _average_ocr_box_confidence(boxes: Iterable[Any]) -> float:
-    scores: list[tuple[float, int]] = []
-    for box in list(boxes or []):
-        try:
-            score = float(getattr(box, "score", 0.0))
-        except (TypeError, ValueError):
-            score = 0.0
-        if score <= 0.0:
-            continue
-        text = str(getattr(box, "text", "") or "")
-        scores.append((score, _ocr_score_weight(text)))
-    if not scores:
-        return 0.0
-    return round(max(0.0, min(_weighted_ocr_score(scores), 1.0)), 3)
-
-
-def _bounded_confidence_or_zero(value: object) -> float:
-    try:
-        return round(max(0.0, min(float(value), 1.0)), 3)
-    except (TypeError, ValueError):
-        return 0.0
 
 
 @dataclass(slots=True)
@@ -1592,31 +1168,6 @@ class DetectedGameWindow:
         }
 
 
-def _matches_aihong_target(target: DetectedGameWindow | None) -> bool:
-    if target is None:
-        return False
-    return _matches_aihong_target_info(
-        process_name=target.process_name,
-        normalized_title=target.normalized_title,
-    )
-
-
-def _builtin_capture_profile_for_target(target: DetectedGameWindow) -> OcrCaptureProfile | None:
-    return _builtin_capture_profile_for_target_stage(target, stage=_AIHONG_DIALOGUE_STAGE)
-
-
-def _builtin_capture_profile_for_target_stage(
-    target: DetectedGameWindow,
-    *,
-    stage: str,
-) -> OcrCaptureProfile | None:
-    if not _matches_aihong_target(target):
-        return None
-    if stage == _AIHONG_MENU_STAGE:
-        return OcrCaptureProfile.from_dict(_AIHONG_MENU_CAPTURE_PROFILE_PRESET)
-    return OcrCaptureProfile.from_dict(_AIHONG_DIALOGUE_CAPTURE_PROFILE_PRESET)
-
-
 @dataclass(slots=True)
 class _StableOcrTextState:
     last_raw_text: str = ""
@@ -1648,213 +1199,6 @@ def _canonical_choice_candidate_text(choices: list[str]) -> str:
 
 def _stripped_ocr_lines(raw_text: str) -> list[str]:
     return [line.strip() for line in str(raw_text or "").splitlines() if line.strip()]
-
-
-@dataclass(slots=True)
-class ParsedOcrCaptureBucket:
-    width: int = 0
-    height: int = 0
-    aspect_ratio: float = 0.0
-    stages: dict[str, OcrCaptureProfile] = field(default_factory=dict)
-
-
-@dataclass(slots=True)
-class ParsedOcrCaptureProcessConfig:
-    stages: dict[str, OcrCaptureProfile] = field(default_factory=dict)
-    window_buckets: dict[str, ParsedOcrCaptureBucket] = field(default_factory=dict)
-
-
-@dataclass(slots=True)
-class ResolvedOcrCaptureSelection:
-    profile: OcrCaptureProfile = field(default_factory=OcrCaptureProfile)
-    match_source: str = OCR_CAPTURE_PROFILE_MATCH_SOURCE_CONFIG_DEFAULT
-    bucket_key: str = ""
-
-
-def _resolve_stage_capture_profile(
-    stage_profiles: dict[str, OcrCaptureProfile],
-    *,
-    stage: str,
-) -> OcrCaptureProfile | None:
-    return stage_profiles.get(stage) or stage_profiles.get(OCR_CAPTURE_PROFILE_STAGE_DEFAULT)
-
-
-def _uses_manual_capture_profile(
-    profiles: dict[str, ParsedOcrCaptureProcessConfig],
-    target: DetectedGameWindow,
-) -> bool:
-    process_name = str(target.process_name or "").strip().lower()
-    if not process_name:
-        return False
-    return process_name in profiles
-
-
-def _lookup_capture_profile(
-    profiles: dict[str, ParsedOcrCaptureProcessConfig],
-    target: DetectedGameWindow,
-    *,
-    stage: str,
-) -> ResolvedOcrCaptureSelection | None:
-    process_name = str(target.process_name or "").strip().lower()
-    if not process_name:
-        return None
-    configured = profiles.get(process_name)
-    if configured is None:
-        return None
-
-    if target.width > 0 and target.height > 0:
-        exact_bucket_key = build_ocr_capture_profile_bucket_key(target.width, target.height).lower()
-        exact_bucket = configured.window_buckets.get(exact_bucket_key)
-        if exact_bucket is not None:
-            exact_profile = _resolve_stage_capture_profile(exact_bucket.stages, stage=stage)
-            if exact_profile is not None:
-                return ResolvedOcrCaptureSelection(
-                    profile=exact_profile,
-                    match_source=OCR_CAPTURE_PROFILE_MATCH_SOURCE_BUCKET_EXACT,
-                    bucket_key=exact_bucket_key,
-                )
-
-        target_aspect_ratio = target.aspect_ratio
-        if target_aspect_ratio > 0:
-            nearest_bucket_key = ""
-            nearest_profile: OcrCaptureProfile | None = None
-            nearest_size_delta: int | None = None
-            nearest_aspect_delta: float | None = None
-            for bucket_key, bucket in configured.window_buckets.items():
-                profile = _resolve_stage_capture_profile(bucket.stages, stage=stage)
-                if profile is None:
-                    continue
-                aspect_delta = abs(float(bucket.aspect_ratio or 0.0) - target_aspect_ratio)
-                if aspect_delta > 0.03:
-                    continue
-                size_delta = abs(int(bucket.width or 0) - target.width) + abs(
-                    int(bucket.height or 0) - target.height
-                )
-                if (
-                    nearest_size_delta is None
-                    or size_delta < nearest_size_delta
-                    or (
-                        size_delta == nearest_size_delta
-                        and (
-                            nearest_aspect_delta is None
-                            or aspect_delta < nearest_aspect_delta
-                        )
-                    )
-                ):
-                    nearest_bucket_key = bucket_key
-                    nearest_profile = profile
-                    nearest_size_delta = size_delta
-                    nearest_aspect_delta = aspect_delta
-            if nearest_profile is not None:
-                return ResolvedOcrCaptureSelection(
-                    profile=nearest_profile,
-                    match_source=OCR_CAPTURE_PROFILE_MATCH_SOURCE_BUCKET_ASPECT_NEAREST,
-                    bucket_key=nearest_bucket_key,
-                )
-
-    fallback_profile = _resolve_stage_capture_profile(configured.stages, stage=stage)
-    if fallback_profile is not None:
-        return ResolvedOcrCaptureSelection(
-            profile=fallback_profile,
-            match_source=OCR_CAPTURE_PROFILE_MATCH_SOURCE_PROCESS_FALLBACK,
-        )
-    return None
-
-
-def _parse_configured_capture_profiles(
-    profiles: dict[str, dict[str, Any]],
-    logger,
-) -> dict[str, ParsedOcrCaptureProcessConfig]:
-    parsed_profiles: dict[str, ParsedOcrCaptureProcessConfig] = {}
-    for process_name, profile_value in profiles.items():
-        normalized_process_name = str(process_name or "").strip().lower()
-        if not normalized_process_name or not isinstance(profile_value, dict):
-            continue
-        stage_profiles: dict[str, OcrCaptureProfile] = {}
-        if all(key in profile_value for key in OCR_CAPTURE_PROFILE_RATIO_KEYS):
-            try:
-                stage_profiles[OCR_CAPTURE_PROFILE_STAGE_DEFAULT] = OcrCaptureProfile.from_dict(
-                    profile_value
-                )
-            except Exception as exc:
-                logger.warning(
-                    "ocr_reader failed to parse capture profile for {}: {}",
-                    normalized_process_name,
-                    exc,
-                )
-                continue
-        else:
-            for stage_name, stage_profile in profile_value.items():
-                normalized_stage_name = str(stage_name or "").strip()
-                if normalized_stage_name == OCR_CAPTURE_PROFILE_WINDOW_BUCKETS_KEY:
-                    continue
-                if not normalized_stage_name or not isinstance(stage_profile, dict):
-                    continue
-                try:
-                    stage_profiles[normalized_stage_name] = OcrCaptureProfile.from_dict(stage_profile)
-                except Exception as exc:
-                    logger.warning(
-                        "ocr_reader failed to parse capture profile for {}/{}: {}",
-                        normalized_process_name,
-                        normalized_stage_name,
-                        exc,
-                    )
-        window_buckets: dict[str, ParsedOcrCaptureBucket] = {}
-        raw_buckets = profile_value.get(OCR_CAPTURE_PROFILE_WINDOW_BUCKETS_KEY)
-        if isinstance(raw_buckets, dict):
-            for bucket_key, bucket_value in raw_buckets.items():
-                normalized_bucket_key = str(bucket_key or "").strip().lower()
-                parsed_dimensions = parse_ocr_capture_profile_bucket_key(normalized_bucket_key)
-                if parsed_dimensions is None or not isinstance(bucket_value, dict):
-                    continue
-                try:
-                    width = int(bucket_value.get("width") or parsed_dimensions[0])
-                    height = int(bucket_value.get("height") or parsed_dimensions[1])
-                except (TypeError, ValueError):
-                    continue
-                if width <= 0 or height <= 0:
-                    continue
-                try:
-                    aspect_ratio = float(
-                        bucket_value.get("aspect_ratio")
-                        or compute_ocr_window_aspect_ratio(width, height)
-                    )
-                except (TypeError, ValueError):
-                    aspect_ratio = compute_ocr_window_aspect_ratio(width, height)
-                raw_stages = bucket_value.get("stages")
-                if not isinstance(raw_stages, dict):
-                    continue
-                bucket_stages: dict[str, OcrCaptureProfile] = {}
-                for stage_name, stage_profile in raw_stages.items():
-                    normalized_stage_name = str(stage_name or "").strip()
-                    if not normalized_stage_name or not isinstance(stage_profile, dict):
-                        continue
-                    try:
-                        bucket_stages[normalized_stage_name] = OcrCaptureProfile.from_dict(
-                            stage_profile
-                        )
-                    except Exception as exc:
-                        logger.warning(
-                            "ocr_reader failed to parse capture profile for {}/{}/{}: {}",
-                            normalized_process_name,
-                            normalized_bucket_key,
-                            normalized_stage_name,
-                            exc,
-                        )
-                if bucket_stages:
-                    canonical_bucket_key = build_ocr_capture_profile_bucket_key(width, height).lower()
-                    window_buckets[canonical_bucket_key] = ParsedOcrCaptureBucket(
-                        width=width,
-                        height=height,
-                        aspect_ratio=aspect_ratio,
-                        stages=bucket_stages,
-                    )
-        if stage_profiles or window_buckets:
-            parsed_profiles[normalized_process_name] = ParsedOcrCaptureProcessConfig(
-                stages=stage_profiles,
-                window_buckets=window_buckets,
-            )
-    return parsed_profiles
 
 
 @dataclass(slots=True)
